Updated documentation . Relocated a Ninject project into the Ninject folder.
