+++ Cqrs-2012.sln
@@ -209,7 +209,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs.Ninject.EventStore", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs.Ninject.Mongo", "Framework\Ninject\Cqrs.Ninject.Mongo\Cqrs.Ninject.Mongo.csproj", "{0955D57A-5B86-4145-92A6-84CEFFA8EC2C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs.Ninject.ServiceHost", "Framework\Cqrs.Ninject.ServiceHost\Cqrs.Ninject.ServiceHost.csproj", "{C2FE26B7-EA82-43C8-ADAC-ED409D50CCF6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs.Ninject.ServiceHost", "Framework\Ninject\Cqrs.Ninject.ServiceHost\Cqrs.Ninject.ServiceHost.csproj", "{C2FE26B7-EA82-43C8-ADAC-ED409D50CCF6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Sample", "Sample", "{3A2BBC96-9AEE-4E71-B97F-4E78F3D446E3}"
 EndProject
