+++ Cqrs.sln
@@ -158,7 +158,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs.Ninject.Azure.ServiceB
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs.Ninject.Azure.ServiceBus.CommandBus", "Framework\Ninject\Azure\Cqrs.Ninject.Azure.ServiceBus.CommandBus\Cqrs.Ninject.Azure.ServiceBus.CommandBus.csproj", "{6DFE1B83-DB0B-414B-83D0-AFB81D9D6397}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs.Ninject.ServiceHost", "Framework\Cqrs.Ninject.ServiceHost\Cqrs.Ninject.ServiceHost.csproj", "{C2FE26B7-EA82-43C8-ADAC-ED409D50CCF6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs.Ninject.ServiceHost", "Framework\Ninject\Cqrs.Ninject.ServiceHost\Cqrs.Ninject.ServiceHost.csproj", "{C2FE26B7-EA82-43C8-ADAC-ED409D50CCF6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs.Modelling.UmlProfiles.VS2010", "Modelling\UmlProfiles\Cqrs.Modelling.UmlProfiles.VS2010.csproj", "{29FAA491-025F-4878-9627-4ADD4E844727}"
 EndProject
