+++ Framework/Azure/Cqrs.Azure.BlobStorage/Factories/BlobStorageDataStoreFactory.cs
@@ -8,11 +8,12 @@
 
 using cdmdotnet.Logging;
 using Cqrs.Azure.BlobStorage.DataStores;
+using Cqrs.DataStores;
 
 namespace Cqrs.Azure.BlobStorage.Factories
 {
 	/// <summary>
-	/// A factory for obtaining DataStore collections from Azure Blob Storage
+	/// A factory for obtaining <see cref="IDataStore{TData}"/> collections from Azure Blob Storage.
 	/// </summary>
 	public abstract class BlobStorageDataStoreFactory
 	{
@@ -20,6 +21,9 @@ namespace Cqrs.Azure.BlobStorage.Factories
 
 		protected ILogger Logger { get; private set; }
 
+		/// <summary>
+		/// Instantiates and Initialises a new instance of the <see cref="BlobStorageDataStoreFactory"/> class.
+		/// </summary>
 		protected BlobStorageDataStoreFactory(ILogger logger, IBlobStorageDataStoreConnectionStringFactory blobStorageDataStoreConnectionStringFactory)
 		{
 			Logger = logger;
