+++ Framework/Azure/Cqrs.Azure.DocumentDb/DataStores/AzureDocumentDbDataStoreExtensions.cs
@@ -1,3 +1,11 @@
+#region Copyright
+// // -----------------------------------------------------------------------
+// // <copyright company="Chinchilla Software Limited">
+// // 	Copyright Chinchilla Software Limited. All rights reserved.
+// // </copyright>
+// // -----------------------------------------------------------------------
+#endregion
+
 using System.Collections.Generic;
 using Cqrs.Azure.DocumentDb.DataStores;
 using Cqrs.Azure.DocumentDb.Entities;
