+++ Framework/Azure/Cqrs.Azure.DocumentDb/Events/AzureDocumentDbEventBuilder.cs
@@ -1,4 +1,12 @@
-﻿using System;
+﻿#region Copyright
+// // -----------------------------------------------------------------------
+// // <copyright company="Chinchilla Software Limited">
+// // 	Copyright Chinchilla Software Limited. All rights reserved.
+// // </copyright>
+// // -----------------------------------------------------------------------
+#endregion
+
+using System;
 using Cqrs.Events;
 
 namespace Cqrs.Azure.DocumentDb.Events
