+++ Framework/Azure/Cqrs.Azure.DocumentDb/Events/AzureDocumentDbEventStoreConnectionStringFactory.cs
@@ -1,4 +1,12 @@
-﻿using System;
+﻿#region Copyright
+// // -----------------------------------------------------------------------
+// // <copyright company="Chinchilla Software Limited">
+// // 	Copyright Chinchilla Software Limited. All rights reserved.
+// // </copyright>
+// // -----------------------------------------------------------------------
+#endregion
+
+using System;
 using cdmdotnet.Logging;
 using Cqrs.Configuration;
 using Microsoft.Azure.Documents.Client;
