+++ Framework/Cqrs.Mongo/Entities/Entity.cs
@@ -9,22 +9,37 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 using System.Runtime.Serialization;
+using Cqrs.DataStores;
+using Cqrs.Entities;
 using MongoDB.Bson.Serialization.Attributes;
 
 namespace Cqrs.Mongo.Entities
 {
+	/// <summary>
+	/// A projection/entity especially designed to work with MongoDB.
+	/// </summary>
 	[Serializable]
 	[DataContract]
-	public abstract class MongoEntity : Cqrs.Entities.Entity
+	public abstract class MongoEntity : Entity
 	{
+		/// <summary>
+		/// The identifier of the <see cref="IEntity"/>.
+		/// </summary>
 		[Required]
 		[BsonId]
 		[DataMember]
 		public override Guid Rsn { get; set; }
 
+		/// <summary>
+		/// The order of this <see cref="IEntity"/> to sort by, by default.
+		/// </summary>
 		[DataMember]
 		public override int SortingOrder { get; set; }
 
+		/// <summary>
+		/// Indicates if this <see cref="IEntity"/> has been deleted, but not removed from the <see cref="IDataStore{TData}"/>,
+		/// this way entities can be retrieved so an un-deleted operation can be triggered.
+		/// </summary>
 		[Required]
 		[DataMember]
 		public override bool IsLogicallyDeleted { get; set; }
