--- Framework/Cqrs.Ninject.ServiceHost/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Ninject" version="3.2.0.0" targetFramework="net40" />
-  <package id="Ninject.Extensions.Wcf" version="3.2.2.0" targetFramework="net40" />
-  <package id="Ninject.Web.Common" version="3.2.0.0" targetFramework="net40" />
-</packages>
