+++ Framework/Cqrs.Sql/DataStores/SqlDataStoreExtensions.cs
@@ -1,4 +1,12 @@
-﻿using System;
+﻿#region Copyright
+// // -----------------------------------------------------------------------
+// // <copyright company="Chinchilla Software Limited">
+// // 	Copyright Chinchilla Software Limited. All rights reserved.
+// // </copyright>
+// // -----------------------------------------------------------------------
+#endregion
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
@@ -6,10 +14,13 @@ using Cqrs.DataStores;
 
 namespace Cqrs.Sql.DataStores
 {
+	/// <summary>
+	/// A collection of extension methods for <see cref="IDataStore{TData}"/>
+	/// </summary>
 	public static class SqlDataStoreExtensions
 	{
 		/// <summary>
-		/// Use this one
+		/// Use this one... Filters a sequence of values based on a predicate.
 		/// </summary>
 		public static IQueryable<TEntity> Where<TEntity, TDbEntity>(this IDataStore<TEntity> dataStore, Expression<Func<TEntity, bool>> predicate)
 			where TDbEntity : class, new()
@@ -32,7 +43,7 @@ namespace Cqrs.Sql.DataStores
 		}
 
 		/// <summary>
-		/// Use this one
+		/// Use this one... Returns the only element of a sequence, and throws an exception if there is not exactly one element in the sequence.
 		/// </summary>
 		public static TEntity Single<TEntity, TDbEntity>(this IEnumerable<TEntity> dataStore)
 			where TDbEntity : class, new()
