+++ Framework/Cqrs/Configuration/ConfigurationManager.cs
@@ -10,10 +10,17 @@ using System;
 
 namespace Cqrs.Configuration
 {
+	/// <summary>
+	/// Provides access to configuration settings from the app settings of an app.config or web.config... i.e. <see cref="System.Configuration.ConfigurationManager.AppSettings"/>
+	/// </summary>
 	public class ConfigurationManager : IConfigurationManager
 	{
 		#region Implementation of IConfigurationManager
 
+		/// <summary>
+		/// Read the setting named <paramref name="key"/>.
+		/// </summary>
+		/// <param name="key">The key (or name) of the setting to read.</param>
 		public virtual string GetSetting(string key)
 		{
 			return System.Configuration.ConfigurationManager.AppSettings[key];
