+++ Framework/Cqrs/Configuration/IConfigurationManager.cs
@@ -8,8 +8,15 @@
 
 namespace Cqrs.Configuration
 {
+	/// <summary>
+	/// Provides access to configuration settings.
+	/// </summary>
 	public interface IConfigurationManager
 	{
+		/// <summary>
+		/// Read the setting named <paramref name="key"/>.
+		/// </summary>
+		/// <param name="key">The key (or name) of the setting to read.</param>
 		string GetSetting(string key);
 
 		bool TryGetSetting(string key, out string value);
