+++ Framework/Cqrs/Entities/DecimalRange.cs
@@ -1,4 +1,10 @@
-﻿using System;
+﻿#region Copyright
+// // -----------------------------------------------------------------------
+// // <copyright company="Chinchilla Software Limited">
+// // 	Copyright Chinchilla Software Limited. All rights reserved.
+// // </copyright>
+// // -----------------------------------------------------------------------
+#endregion
 
 namespace Cqrs.Entities
 {
