+++ Framework/Ninject/Cqrs.Ninject.ServiceHost/Cqrs.Ninject.ServiceHost.nuspec
@@ -1,0 +1,27 @@
+<?xml version="1.0"?>
+<package>
+	<metadata>
+		<id>Cqrs.Ninject.ServiceHost</id>
+		<title>Microsoft WCF with CQRS.NET and Ninject ServiceHosts</title>
+		<version>$version$</version>
+		<authors>Chinchilla Software</authors>
+		<licenseUrl>http://www.apache.org/licenses/LICENSE-2.0</licenseUrl>
+		<projectUrl>http://chinchilla-software-com.github.io/CQRS/</projectUrl>
+		<iconUrl>http://chinchilla-software-com.github.io/CQRS/wiki/ChinChilla-Software-Red.png</iconUrl>
+		<requireLicenseAcceptance>false</requireLicenseAcceptance>
+		<description>Use Ninject as your IoC container of choice to use WCF with CQRS.NET as ServiceHosts</description>
+		<copyright>Copyright 2013</copyright>
+		<tags>CQRS CQRS.NET Pub/Sub Microservices ServiceBus Ninject</tags>
+		<releaseNotes>
+			Version 2.0
+
+			* New
+		</releaseNotes>
+		<dependencies>
+			<dependency id="Cqrs" version="[$version$]" />
+		</dependencies>
+	</metadata>
+	<files>
+		<file src="Tools\*.ps1" target="Tools" />
+	</files>
+</package>
