+++ Framework/Ninject/Cqrs.Ninject.ServiceHost/Tools/install.ps1
