--- Framework/Web/Cqrs.Web.AspNet.WebApi/Class1.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Cqrs.Web.AspNet.WebApi
-{
-    public class Class1
-    {
-    }
-}
