+++ packages/repositories.config
@@ -40,6 +40,7 @@
   <repository path="..\Framework\Ninject\Cqrs.Ninject.InProcess.EventStore\packages.config" />
   <repository path="..\Framework\Ninject\Cqrs.Ninject.Mongo\packages.config" />
   <repository path="..\Framework\Ninject\Cqrs.Ninject.MongoDB\packages.config" />
+  <repository path="..\Framework\Ninject\Cqrs.Ninject.ServiceHost\packages.config" />
   <repository path="..\Framework\Ninject\Cqrs.Ninject.WebApi\packages.config" />
   <repository path="..\Framework\Ninject\Cqrs.Ninject\packages.config" />
   <repository path="..\HelloWorldExample\packages.config" />
